requirements.txt
django
djangorestframework
django-cors-headers
django-multiselectfield
Pillow
